% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion.R
\name{fd_accordion_item}
\alias{fd_accordion_item}
\title{Accordion Item for Foundation UI}
\usage{
fd_accordion_item(title, content, is_active = FALSE)
}
\arguments{
\item{title}{Title of the item, will be seen when collapsed.}

\item{content}{Shiny tag containing the UI when the item is opened.}

\item{is_active}{Logical, should the item be open on start-up?}
}
\description{
An individual item for the accordion.
}
\seealso{
\code{\link{fd_accordion}}
}
