% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{fd_action_button}
\alias{fd_action_button}
\alias{fd_link_button}
\title{Button for Foundation UI}
\source{
\url{https://foundation.zurb.com/sites/docs/button.html}
}
\usage{
fd_action_button(id, label, class = NULL, disabled = FALSE, ...)

fd_link_button(id, label, link, class = NULL, ...)
}
\arguments{
\item{id}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link.}

\item{class}{Additional styles to add to the class. This can include (but not limited to):
\describe{
\item{\code{expanded}}{The button takes up the full available width.}
\item{\code{tiny}, \code{small}, \code{large}}{Changes the size of the button.}
\item{\code{primary}, \code{secondary}, \code{success}, \code{alert}, \code{warning}}{Changes the colour of the button.}
\item{\code{hollow}}{Turns the button to a hollow/non-block colour style.}
\item{\code{clear}}{Turns the button to a clear style, almost like a link with padding.}
}}

\item{disabled}{Logical. Should the button be disabled?}

\item{...}{Named attributes to be applied to the button}

\item{link}{The URL/path to direct the user to.}
}
\description{
Creates a button that either is a link or performs an action when clicked.
}
