% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callout.R
\name{fd_callout}
\alias{fd_callout}
\title{Callout for Foundation UI}
\usage{
fd_callout(..., class = "", closable = FALSE)
}
\arguments{
\item{...}{tags of the contents of the callout.}

\item{class}{Option to add the type of}
}
\description{
A
}
\examples{
fd_callout(
  tags$p("This is an example callout")
)

}
