% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{fd_helptext}
\alias{fd_helptext}
\title{Help Text for Foundation Inputs}
\usage{
fd_helptext(...)
}
\description{
A piece of text that can be placed underneath an input that can be used
to describe what is required for the input.
}
