% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{fd_numeric}
\alias{fd_numeric}
\title{Numeric Input for Foundation UI}
\usage{
fd_numeric(id, label, value, min = NULL, max = NULL, step = NULL)
}
\description{
Numeric Input for Foundation UI
}
\examples{
if (interactive()) {
  ui <- fd_page(
    fd_numeric("test_num", "Number", 5, min = 1, max = 10),
    shiny::textOutput("out_num")
  )

  server <- function(input, output) {
    output$out_num <- shiny::renderText(input$test_num)
  }
}

}
