% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundationPage.R
\name{foundationPage}
\alias{foundationPage}
\title{Basic Page for Foundation UI}
\usage{
foundationPage(...)
}
\arguments{
\item{...}{The contents of the document body.}
}
\description{
Creates a Shiny UI page that load s the CSS and JavaScript for Foundation.
}
