% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundationObjects.R
\name{fd_grid}
\alias{fd_grid}
\title{Foundation UI Grid}
\source{
\url{https://foundation.zurb.com/sites/docs/xy-grid.html}
}
\usage{
fd_grid(..., direction = "x", margin = "", padding = "",
  small_max = NULL, medium_max = NULL, large_max = NULL)
}
\arguments{
\item{...}{Elements to include within the grid}

\item{direction}{Which direction should the elements go, can either be \code{x} or \code{y}}

\item{margin, padding}{Which direction}

\item{small_max, medium_max, large_max}{The number of cells in each column/row, depending on what
\code{direction} it is. The default will be that a cell takes up the full width/height.}
}
\description{
The UI grid is based off Foundation's XY-Grid. This allows the grid to either be set by columns or rows.
}
\examples{
fd_container(
  fd_grid(
    fd_cell("Cell 1"),
    fd_cell("Cell 2"),
    fd_cell("Cell 3")
  )
)

}
