% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{fd_page}
\alias{fd_page}
\title{Basic Page for Foundation UI}
\usage{
fd_page(...)
}
\arguments{
\item{...}{The contents of the document body.}
}
\description{
Creates a Shiny UI page that load s the CSS and JavaScript for Foundation.
}
